backend/requirements.txt
Flask==1.0.2
Flask-RESTful==0.3.6
Flask-Cors==3.0.6
pyserial==3.4
peewee==3.14.8
reportlab==3.5.19
pyyaml==3.13
python-escpos==2.2.0
marshmallow==3.14.1
marshmallow-enum==1.5.1
marshmallow-dataclass==8.5.3
SQLAlchemy==1.3.8
marshmallow-sqlalchemy==0.26.1